CCPMA-1068 [V] Accordion (component) (Aurora)
add a flat to ListRow
remove PureListRow
