+++ .gitignore
@@ -5,7 +5,6 @@
 
 # Build directories
 .next
-dist
 coverage
 catalog/build
