+++ src/components/List/index.js
@@ -1,6 +1,5 @@
 export { default as ListContainer } from "./Container";
 export { default as ListRow } from "./Row";
-export { default as PureListRow } from "./PureRow";
 export { default as ListUnstyled } from "./Unstyled.styles";
 export { default as ListRowOverflow } from "./Overflow";
 export { default as Section } from "./Section";
