+++ src/index.js
@@ -22,7 +22,6 @@ export { default as FeedbackInline } from "./components/FeedbackInline";
 export {
   ListContainer,
   ListRow,
-  PureListRow,
   ListRowOverflow,
   Section,
   SectionItem
